+++ src/main/java/net/goldorion/toasts/ToastGUI.java
@@ -1,34 +1,34 @@
 package net.goldorion.toasts;
 
 import net.mcreator.ui.MCreator;
-import net.mcreator.ui.component.util.ComboBoxUtil;
+import net.mcreator.ui.component.JColor;
 import net.mcreator.ui.component.util.ComponentUtils;
 import net.mcreator.ui.component.util.PanelUtils;
 import net.mcreator.ui.help.HelpUtils;
 import net.mcreator.ui.init.L10N;
-import net.mcreator.ui.laf.renderer.WTextureComboBoxRenderer;
+import net.mcreator.ui.minecraft.TextureComboBox;
 import net.mcreator.ui.modgui.ModElementGUI;
-import net.mcreator.ui.validation.AggregatedValidationResult;
 import net.mcreator.ui.validation.ValidationGroup;
 import net.mcreator.ui.validation.component.VTextField;
 import net.mcreator.ui.validation.validators.TextFieldValidator;
 import net.mcreator.ui.workspace.resources.TextureType;
-import net.mcreator.util.ListUtils;
 import net.mcreator.workspace.elements.ModElement;
 
 import javax.annotation.Nonnull;
 import javax.swing.*;
 import java.awt.*;
-import java.io.File;
-import java.util.Collections;
+import java.net.URI;
 
 public class ToastGUI extends ModElementGUI<Toast> {
 
     private final VTextField title = new VTextField(12);
+    private final JColor titleColor = new JColor(mcreator, true, false);
     private final VTextField description = new VTextField(30);
+    private final JColor descriptionColor = new JColor(mcreator, true, false);
+    private final JCheckBox drawShadows = L10N.checkbox("elementgui.common.enable");
     private final JSpinner durationTime = new JSpinner(new SpinnerNumberModel(5000, 1000, 600000, 500));
-    private final JComboBox<String> icon = new JComboBox<>();
-    private final JComboBox<String> background = new JComboBox<>();
+    private final TextureComboBox icon = new TextureComboBox(mcreator, TextureType.SCREEN);
+    private final TextureComboBox background = new TextureComboBox(mcreator, TextureType.SCREEN);
 
     private final ValidationGroup page1group = new ValidationGroup();
 
@@ -43,44 +43,49 @@ public class ToastGUI extends ModElementGUI<Toast> {
         ComponentUtils.deriveFont(title, 16.0F);
         ComponentUtils.deriveFont(description, 16.0F);
 
-        icon.setPrototypeDisplayValue("XXXXXXXXXXXXXXXXXXXXXXXXXX");
-        icon.setRenderer(new WTextureComboBoxRenderer.TypeTextures(mcreator.getWorkspace(), TextureType.SCREEN));
-        background.setPrototypeDisplayValue("XXXXXXXXXXXXXXXXXXXXXXXXXX");
-        background.setRenderer(new WTextureComboBoxRenderer.TypeTextures(mcreator.getWorkspace(), TextureType.SCREEN));
-
-        JPanel mainPanel = new JPanel(new GridLayout(5, 2, 5, 5));
+        JPanel mainPanel = new JPanel(new GridLayout(8, 2, 5, 5));
         mainPanel.setOpaque(false);
 
         mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/title"), L10N.label("elementgui.toast.title")));
         mainPanel.add(title);
-        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/description"), L10N.label("elementgui.toast.description")));
-        mainPanel.add(description);
-        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/duration_time"), L10N.label("elementgui.toast.duration_time")));
-        mainPanel.add(durationTime);
-        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/icon"), L10N.label("elementgui.toast.icon")));
-        mainPanel.add(icon);
-        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/background"), L10N.label("elementgui.toast.background")));
-        mainPanel.add(background);
-
         page1group.addValidationElement(title);
         title.setValidator(new TextFieldValidator(title, L10N.t("elementgui.toast.cant_be_empty", title)));
         title.enableRealtimeValidation();
 
-        addPage(L10N.t("elementgui.common.page_mainPanel"), PanelUtils.totalCenterInPanel(mainPanel));
-    }
+        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/title_color"), L10N.label("elementgui.toast.title_color")));
+        mainPanel.add(titleColor);
+        titleColor.setOpaque(false);
 
-    public void reloadDataLists() {
-        ComboBoxUtil.updateComboBoxContents(icon, ListUtils.merge(Collections.singleton("None"),
-                mcreator.getFolderManager().getTexturesList(TextureType.SCREEN).stream()
-                        .map(File::getName).toList()), "None");
-        ComboBoxUtil.updateComboBoxContents(background, ListUtils.merge(Collections.singleton("Default"),
-                mcreator.getFolderManager().getTexturesList(TextureType.SCREEN).stream()
-                        .map(File::getName).toList()), "Default");
+        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/description"), L10N.label("elementgui.toast.description")));
+        mainPanel.add(description);
+
+        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/description_color"), L10N.label("elementgui.toast.description_color")));
+        mainPanel.add(descriptionColor);
+        descriptionColor.setOpaque(false);
+
+        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/draw_shadows"), L10N.label("elementgui.toast.draw_shadows")));
+        mainPanel.add(drawShadows);
+        drawShadows.setOpaque(false);
+
+        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/duration_time"), L10N.label("elementgui.toast.duration_time")));
+        mainPanel.add(durationTime);
+
+        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/icon"), L10N.label("elementgui.toast.icon")));
+        mainPanel.add(icon);
+
+        mainPanel.add(HelpUtils.wrapWithHelpButton(this.withEntry("toast/background"), L10N.label("elementgui.toast.background")));
+        mainPanel.add(background);
+
+
+        addPage(L10N.t("elementgui.common.page_mainPanel"), PanelUtils.totalCenterInPanel(mainPanel)).validate(page1group);
     }
 
     @Override
-    protected AggregatedValidationResult validatePage(int i) {
-        return new AggregatedValidationResult(page1group);
+    public void reloadDataLists() {
+        super.reloadDataLists();
+
+        icon.reload();
+        background.reload();
     }
 
     @Override
@@ -88,10 +93,11 @@ public class ToastGUI extends ModElementGUI<Toast> {
         title.setText(toast.title);
         description.setText(toast.description);
         durationTime.setValue(toast.durationTime);
-        icon.setSelectedItem(toast.icon);
-        background.setSelectedItem(toast.background);
-        if (toast.background == null)
-            background.setSelectedItem("Default");
+        icon.setTextureFromTextureName(toast.icon);
+        background.setTextureFromTextureName(toast.background);
+        titleColor.setColor(toast.titleColor);
+        descriptionColor.setColor(toast.descriptionColor);
+        drawShadows.setSelected(toast.drawShadows);
     }
 
     @Override
@@ -100,9 +106,17 @@ public class ToastGUI extends ModElementGUI<Toast> {
         toast.title = title.getText();
         toast.description = description.getText();
         toast.durationTime = (int) durationTime.getValue();
-        toast.icon = (String) icon.getSelectedItem();
-        toast.background = (String) background.getSelectedItem();
+        toast.icon = icon.getTexture().getTextureName();
+        toast.background = background.getTexture().getTextureName();
+        toast.titleColor = titleColor.getColor();
+        toast.descriptionColor = descriptionColor.getColor();
+        toast.drawShadows = drawShadows.isSelected();
 
         return toast;
+    }
+
+    @Override
+    public URI contextURL() {
+        return null;
     }
 }