+++ src/main/java/net/goldorion/toasts/ToastsLauncher.java
@@ -1,7 +1,6 @@
 package net.goldorion.toasts;
 
 import net.mcreator.element.ModElementType;
-import net.mcreator.element.ModElementTypeLoader;
 import net.mcreator.plugin.JavaPlugin;
 import net.mcreator.plugin.Plugin;
 import net.mcreator.plugin.events.PreGeneratorsLoadingEvent;
@@ -13,10 +12,11 @@ import static net.mcreator.element.ModElementTypeLoader.register;
 public class ToastsLauncher extends JavaPlugin {
 
     private static final Logger LOG = LogManager.getLogger("Toasts");
+
     public ToastsLauncher(Plugin plugin) {
         super(plugin);
 
-        addListener(PreGeneratorsLoadingEvent.class, e-> register(new ModElementType<>("toast", 'T', ToastGUI::new, Toast.class)));
+        addListener(PreGeneratorsLoadingEvent.class, e -> register(new ModElementType<>("toast", 'T', ToastGUI::new, Toast.class)));
 
         LOG.info("Toasts plugin was loaded");
     }