+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2
+- Updated to MCreator 2025.2
+- Removed support for all previous Minecraft versions
+- Added support for NF 1.21.1 and NF 1.21.4
+- Added new parameters: title's color, description's color and Draw shadows
+- Added 2 new procedure blocks: Show toast with time, Show toast with time and custom icon
+- Fixed some rendering problems
+
 ## 1.1
 - Added a parameter to use a custom background texture
 - Some minor changes